requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
autoawq==0.2.9
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
bitsandbytes==0.48.1
blake3==1.0.7
blinker==1.4
cachetools==6.2.0
cbor2==5.7.0
certifi==2025.8.3
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.3
click==8.3.0
cloud-init==25.1.4
cloudpickle==3.1.1
colorama==0.4.4
command-not-found==0.3
compressed-tensors==0.11.0
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cupy-cuda12x==13.6.0
cut-cross-entropy==25.1.1
datasets==3.6.0
dbus-python==1.2.18
depyf==0.19.0
diffusers==0.35.1
dill==0.3.8
diskcache==5.6.3
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
fastrlock==0.8.3
filelock==3.19.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
gguf==0.17.1
h11==0.16.0
hf-xet==1.1.10
hf_transfer==0.1.9
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.5
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
incremental==21.3.0
interegular==0.3.3
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==23.5.0
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistral_common==1.8.5
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
netifaces==0.11.0
networkx==3.4.2
ninja==1.13.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
openai==2.1.0
openai-harmony==0.0.4
opencv-python-headless==4.12.0.88
outlines_core==0.2.11
packaging==25.0
pandas==2.3.2
partial-json-parser==0.2.1.1.post6
peft==0.17.1
pillow==11.3.0
postal==1.1.10
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pycurl==7.44.1
pydantic==2.11.10
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
Pygments==2.19.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
ray==2.49.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
sentencepiece==0.2.1
sentry-sdk==2.39.0
service-identity==18.1.0
setproctitle==1.3.7
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.48.0
sympy==1.14.0
systemd-python==234
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
torchao==0.13.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
trl==0.23.0
Twisted==22.1.0
typeguard==4.4.4
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.9.31
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unsloth==2025.9.11
unsloth_zoo==2025.9.14
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
vllm==0.11.0
wadllib==1.3.6
watchfiles==1.1.0
websockets==15.0.1
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.5.0
yarl==1.20.1
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.25.0
